svn merge -r 1065:2309 http://code.taobao.org/svn/tfs/branches/tfs-tmp-trunk
git-svn-id: http://code.taobao.org/svn/tfs@2312 b722c2b5-6a69-4643-bcd1-1a75825fa998

+++ trunk/src/message/client_cmd_message.h
@@ -22,7 +22,7 @@ namespace tfs
   namespace message
   {
     // Client Command
-    class ClientCmdMessage: public common::BasePacket 
+    class ClientCmdMessage: public common::BasePacket
     {
     public:
       ClientCmdMessage();
@@ -30,11 +30,11 @@ namespace tfs
       virtual int serialize(common::Stream& output) const ;
       virtual int deserialize(common::Stream& input);
       virtual int64_t length() const;
-      inline void set_value1(const uint64_t value)
+      inline void set_value1(const int64_t value)
       {
         info_.value1_ = value;
       }
-      inline uint64_t get_value1() const
+      inline int64_t get_value1() const
       {
         return info_.value1_;
       }
@@ -42,23 +42,23 @@ namespace tfs
       {
         info_.value2_ = value;
       }
-      inline uint64_t get_value2() const
+      inline int64_t get_value2() const
       {
         return info_.value2_;
       }
-      inline void set_value3(const uint32_t value)
+      inline void set_value3(const uint64_t value)
       {
         info_.value3_ = value;
       }
-      inline uint32_t get_value3() const
+      inline uint64_t get_value3() const
       {
         return info_.value3_;
       }
-      inline void set_value4(const uint32_t value)
+      inline void set_value4(const int64_t value)
       {
         info_.value4_ = value;
       }
-      inline uint32_t get_value4() const
+      inline int64_t get_value4() const
       {
         return info_.value4_;
       }
