+++ trunk/src/message/server_status_message.h
@@ -21,7 +21,7 @@ namespace tfs
 {
   namespace message
   {
-    class GetServerStatusMessage: public common::BasePacket 
+    class GetServerStatusMessage: public common::BasePacket
     {
       public:
         GetServerStatusMessage();
@@ -37,7 +37,7 @@ namespace tfs
         {
           return status_type_;
         }
-        inline void set_from_row(const int32_t row)
+        inline void set_from_row(const int64_t row)
         {
           from_row_ = row;
         }
@@ -59,7 +59,7 @@ namespace tfs
         int32_t return_row_;
     };
 
-    class AccessStatInfoMessage: public common::BasePacket 
+    class AccessStatInfoMessage: public common::BasePacket
     {
       public:
         typedef __gnu_cxx ::hash_map<uint32_t, common::Throughput> COUNTER_TYPE;
@@ -126,7 +126,7 @@ namespace tfs
         mutable int32_t has_next_;
     };
 
-    class ShowServerInformationMessage : public common::BasePacket 
+    class ShowServerInformationMessage : public common::BasePacket
     {
     public:
       ShowServerInformationMessage();
